WIP File-manager restructuration
