+++ blocks-parsers/model-parser.js
@@ -1,0 +1,21 @@
+
+/**
+ * The extension of the file that can be loaded using this parser
+ */
+module.exports.FILE_EXTENSIONS = ["txt"];
+
+/**
+ * Receives an object containing all blocks and their respective data, must convert it to a string that will be written in a file
+ * TODO: Block object desc
+ */
+module.exports.convertToString = (blocks) => {
+  return "";
+};
+
+/**
+ * Receives a string in the format specified in the toString function, must recreate the object that was received in the toString function
+ * TODO: Block object desc
+ */
+module.exports.parse = (str) => {
+  return {};
+};
